# Performance optimization for "AddReference" handling