--- a/benchmarks/LetsTrace.Benchmarks/LetsTrace.Benchmarks.csproj
+++ b/benchmarks/LetsTrace.Benchmarks/LetsTrace.Benchmarks.csproj
@@ -1,0 +1,16 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp2.1</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="BenchmarkDotNet" Version="0.10.13" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\LetsTrace\LetsTrace.csproj" />
+  </ItemGroup>
+
+</Project>