--- a/benchmarks/LetsTrace.Benchmarks/LetsTrace.Benchmarks.csproj
+++ b/benchmarks/LetsTrace.Benchmarks/LetsTrace.Benchmarks.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp2.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>