--- a/src/LetsTrace/SpanBuilder.cs
+++ b/src/LetsTrace/SpanBuilder.cs
@@ -14,9 +14,9 @@ namespace LetsTrace
         private readonly string _operationName;
         private readonly ISampler _sampler;
         private readonly IMetrics _metrics;
-        private readonly List<Reference> _references;
         private readonly Dictionary<string, Field> _tags;
 
+        private List<Reference> _references;
         private bool _ignoreActiveSpan;
         private DateTimeOffset? _startTimestamp;
 
@@ -26,7 +26,8 @@ namespace LetsTrace
             _operationName = operationName ?? throw new ArgumentNullException(nameof(operationName));
             _sampler = sampler ?? throw new ArgumentNullException(nameof(sampler));
             _metrics = metrics ?? throw new ArgumentNullException(nameof(metrics));
-            _references = new List<Reference>();
+
+            // There will be tags in most cases so it should be fine to always initiate this variable.
             _tags = new Dictionary<string, Field>();
         }
 
@@ -34,6 +35,14 @@ namespace LetsTrace
         {
             if (referencedContext != null)
             {
+                if (_references == null)
+                {
+                    // Adding an item to a new list normally creates an array with 4 items.
+                    // Since we'll either have zero or one reference in 99% of all cases, we can optimize this
+                    // and provide better performance for zero/one references with slightly worse perf for multiple references.
+                    _references = new List<Reference>(1);
+                }
+
                 _references.Add(new Reference(referenceType, referencedContext));
             }
             return this;
@@ -56,22 +65,21 @@ namespace LetsTrace
 
         public ISpan Start()
         {
-            if (!_ignoreActiveSpan && _references.Count == 0)
+            if (!_ignoreActiveSpan && _references == null)
             {
-                // Note: Stores the variable to ensure the underlying AsyncLocal is accessed only once!
-                ISpan activeSpan = _tracer.ActiveSpan;
-                if (activeSpan != null)
-                {
-                    AsChildOf(activeSpan);
-                }
+                AsChildOf(_tracer.ActiveSpan);
             }
 
             SpanContext parent = null;
-            foreach(var reference in _references)
+            if (_references != null)
             {
-                if (reference.Type == References.ChildOf) {
-                    parent = reference.Context as SpanContext;
-                    break;
+                foreach (var reference in _references)
+                {
+                    if (reference.Type == References.ChildOf)
+                    {
+                        parent = reference.Context as SpanContext;
+                        break;
+                    }
                 }
             }