This PR optimizes the handling of span references:
* No allocation required if the span has no reference
* Less allocations for when the span has one reference
* Downside: Slightly worse performance for when the span has multiple references, but this is a rare case anyway.

I also added a BenchmarkDotNet test project which should be useful in the future and I compared my changes:

#### Before
``` ini
BenchmarkDotNet=v0.10.13, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.309)
Intel Core i5-4300U CPU 1.90GHz (Haswell), 1 CPU, 4 logical cores and 2 physical cores
Frequency=2435766 Hz, Resolution=410.5485 ns, Timer=TSC
.NET Core SDK=2.1.300-preview1-008174
  [Host]     : .NET Core 2.1.0-preview1-26216-03 (CoreCLR 4.6.26216.04, CoreFX 4.6.26216.02), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.0-preview1-26216-03 (CoreCLR 4.6.26216.04, CoreFX 4.6.26216.02), 64bit RyuJIT
```
|               Method |     Mean |     Error |    StdDev |  Gen 0 | Allocated |
|--------------------- |---------:|----------:|----------:|-------:|----------:|
|   Start_No_Reference | 843.7 ns | 11.827 ns | 11.063 ns | 0.5283 |     832 B |
|  Start_One_Reference | 586.3 ns | 10.410 ns |  9.228 ns | 0.4063 |     640 B |
| Start_Two_References | 593.2 ns |  9.583 ns |  8.964 ns | 0.4063 |     640 B |

#### After 
``` ini
BenchmarkDotNet=v0.10.13, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.309)
Intel Core i5-4300U CPU 1.90GHz (Haswell), 1 CPU, 4 logical cores and 2 physical cores
Frequency=2435766 Hz, Resolution=410.5485 ns, Timer=TSC
.NET Core SDK=2.1.300-preview1-008174
  [Host]     : .NET Core 2.1.0-preview1-26216-03 (CoreCLR 4.6.26216.04, CoreFX 4.6.26216.02), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.0-preview1-26216-03 (CoreCLR 4.6.26216.04, CoreFX 4.6.26216.02), 64bit RyuJIT
```
|               Method |     Mean |    Error |    StdDev |  Gen 0 | Allocated |
|--------------------- |---------:|---------:|----------:|-------:|----------:|
|   Start_No_Reference | 808.9 ns | 10.39 ns |  9.715 ns | 0.5026 |     792 B |
|  Start_One_Reference | 560.0 ns | 10.51 ns |  9.318 ns | 0.3757 |     592 B |
| Start_Two_References | 628.4 ns | 11.70 ns | 11.488 ns | 0.4110 |     648 B |

This is similar to how it is implemented [jaeger-client-java](https://github.com/jaegertracing/jaeger-client-java/blob/d69c2d3333da79d7eb1ffdf982186de5b5f1f4bc/jaeger-core/src/main/java/com/uber/jaeger/Tracer.java#L248)